+++ Torrentia/DetailsForm.Designer.cs
@@ -30,13 +30,13 @@
         {
             this.listResults = new System.Windows.Forms.ListView();
             this.columnHeader2 = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
-            this.columnHeader3 = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
             this.txtContents = new System.Windows.Forms.RichTextBox();
             this.panel1 = new System.Windows.Forms.Panel();
             this.lblTorrent = new System.Windows.Forms.Label();
             this.label1 = new System.Windows.Forms.Label();
-            this.button1 = new System.Windows.Forms.Button();
-            this.button2 = new System.Windows.Forms.Button();
+            this.lblHash = new System.Windows.Forms.Label();
+            this.btnDownload = new System.Windows.Forms.Button();
+            this.btnCopyHash = new System.Windows.Forms.Button();
             this.panel1.SuspendLayout();
             this.SuspendLayout();
             // 
@@ -48,15 +48,14 @@
             this.listResults.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(20)))), ((int)(((byte)(20)))), ((int)(((byte)(20)))));
             this.listResults.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.listResults.Columns.AddRange(new System.Windows.Forms.ColumnHeader[] {
-            this.columnHeader2,
-            this.columnHeader3});
+            this.columnHeader2});
             this.listResults.ForeColor = System.Drawing.Color.White;
             this.listResults.FullRowSelect = true;
             this.listResults.Location = new System.Drawing.Point(9, 34);
             this.listResults.MultiSelect = false;
             this.listResults.Name = "listResults";
             this.listResults.ShowGroups = false;
-            this.listResults.Size = new System.Drawing.Size(856, 352);
+            this.listResults.Size = new System.Drawing.Size(856, 328);
             this.listResults.TabIndex = 76;
             this.listResults.UseCompatibleStateImageBehavior = false;
             this.listResults.View = System.Windows.Forms.View.Details;
@@ -66,11 +65,6 @@
             // 
             this.columnHeader2.Text = "Tracker";
             this.columnHeader2.Width = 507;
-            // 
-            // columnHeader3
-            // 
-            this.columnHeader3.Text = "Age";
-            this.columnHeader3.Width = 122;
             // 
             // txtContents
             // 
@@ -91,7 +85,7 @@
             | System.Windows.Forms.AnchorStyles.Right)));
             this.panel1.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.panel1.Controls.Add(this.txtContents);
-            this.panel1.Location = new System.Drawing.Point(9, 432);
+            this.panel1.Location = new System.Drawing.Point(9, 450);
             this.panel1.Name = "panel1";
             this.panel1.Size = new System.Drawing.Size(856, 160);
             this.panel1.TabIndex = 78;
@@ -113,57 +107,71 @@
             this.label1.AutoSize = true;
             this.label1.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.label1.ForeColor = System.Drawing.Color.DodgerBlue;
-            this.label1.Location = new System.Drawing.Point(4, 401);
+            this.label1.Location = new System.Drawing.Point(4, 419);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(94, 28);
+            this.label1.Size = new System.Drawing.Size(170, 28);
             this.label1.TabIndex = 80;
             this.label1.Tag = "themeable";
-            this.label1.Text = "Contents";
+            this.label1.Text = "Torrent Contents:";
             // 
-            // button1
+            // lblHash
             // 
-            this.button1.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
-            this.button1.BackColor = System.Drawing.Color.DodgerBlue;
-            this.button1.FlatAppearance.BorderColor = System.Drawing.Color.DodgerBlue;
-            this.button1.FlatAppearance.MouseDownBackColor = System.Drawing.Color.RoyalBlue;
-            this.button1.FlatAppearance.MouseOverBackColor = System.Drawing.Color.RoyalBlue;
-            this.button1.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
-            this.button1.ForeColor = System.Drawing.Color.White;
-            this.button1.Location = new System.Drawing.Point(710, 392);
-            this.button1.Name = "button1";
-            this.button1.Size = new System.Drawing.Size(155, 34);
-            this.button1.TabIndex = 81;
-            this.button1.Tag = "themeable";
-            this.button1.Text = "Open torrent link";
-            this.button1.UseVisualStyleBackColor = false;
-            this.button1.Click += new System.EventHandler(this.button1_Click);
+            this.lblHash.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.lblHash.AutoSize = true;
+            this.lblHash.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.lblHash.ForeColor = System.Drawing.Color.DodgerBlue;
+            this.lblHash.Location = new System.Drawing.Point(4, 365);
+            this.lblHash.Name = "lblHash";
+            this.lblHash.Size = new System.Drawing.Size(69, 28);
+            this.lblHash.TabIndex = 81;
+            this.lblHash.Tag = "themeable";
+            this.lblHash.Text = "Hash: ";
             // 
-            // button2
+            // btnDownload
             // 
-            this.button2.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
-            this.button2.BackColor = System.Drawing.Color.DodgerBlue;
-            this.button2.FlatAppearance.BorderColor = System.Drawing.Color.DodgerBlue;
-            this.button2.FlatAppearance.MouseDownBackColor = System.Drawing.Color.RoyalBlue;
-            this.button2.FlatAppearance.MouseOverBackColor = System.Drawing.Color.RoyalBlue;
-            this.button2.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
-            this.button2.ForeColor = System.Drawing.Color.White;
-            this.button2.Location = new System.Drawing.Point(549, 392);
-            this.button2.Name = "button2";
-            this.button2.Size = new System.Drawing.Size(155, 34);
-            this.button2.TabIndex = 82;
-            this.button2.Tag = "themeable";
-            this.button2.Text = "Copy magnet URI";
-            this.button2.UseVisualStyleBackColor = false;
-            this.button2.Click += new System.EventHandler(this.button2_Click);
+            this.btnDownload.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.btnDownload.BackColor = System.Drawing.Color.DodgerBlue;
+            this.btnDownload.FlatAppearance.BorderColor = System.Drawing.Color.DodgerBlue;
+            this.btnDownload.FlatAppearance.MouseDownBackColor = System.Drawing.Color.RoyalBlue;
+            this.btnDownload.FlatAppearance.MouseOverBackColor = System.Drawing.Color.RoyalBlue;
+            this.btnDownload.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
+            this.btnDownload.ForeColor = System.Drawing.Color.White;
+            this.btnDownload.Location = new System.Drawing.Point(744, 368);
+            this.btnDownload.Name = "btnDownload";
+            this.btnDownload.Size = new System.Drawing.Size(121, 34);
+            this.btnDownload.TabIndex = 82;
+            this.btnDownload.Tag = "themeable";
+            this.btnDownload.Text = "Download";
+            this.btnDownload.UseVisualStyleBackColor = false;
+            this.btnDownload.Click += new System.EventHandler(this.btnDownload_Click);
+            // 
+            // btnCopyHash
+            // 
+            this.btnCopyHash.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.btnCopyHash.BackColor = System.Drawing.Color.DodgerBlue;
+            this.btnCopyHash.FlatAppearance.BorderColor = System.Drawing.Color.DodgerBlue;
+            this.btnCopyHash.FlatAppearance.MouseDownBackColor = System.Drawing.Color.RoyalBlue;
+            this.btnCopyHash.FlatAppearance.MouseOverBackColor = System.Drawing.Color.RoyalBlue;
+            this.btnCopyHash.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
+            this.btnCopyHash.ForeColor = System.Drawing.Color.White;
+            this.btnCopyHash.Location = new System.Drawing.Point(617, 368);
+            this.btnCopyHash.Name = "btnCopyHash";
+            this.btnCopyHash.Size = new System.Drawing.Size(121, 34);
+            this.btnCopyHash.TabIndex = 83;
+            this.btnCopyHash.Tag = "themeable";
+            this.btnCopyHash.Text = "Copy Hash";
+            this.btnCopyHash.UseVisualStyleBackColor = false;
+            this.btnCopyHash.Click += new System.EventHandler(this.btnCopyHash_Click);
             // 
             // DetailsForm
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(120F, 120F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Dpi;
             this.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(20)))), ((int)(((byte)(20)))), ((int)(((byte)(20)))));
-            this.ClientSize = new System.Drawing.Size(876, 604);
-            this.Controls.Add(this.button2);
-            this.Controls.Add(this.button1);
+            this.ClientSize = new System.Drawing.Size(876, 622);
+            this.Controls.Add(this.btnCopyHash);
+            this.Controls.Add(this.btnDownload);
+            this.Controls.Add(this.lblHash);
             this.Controls.Add(this.label1);
             this.Controls.Add(this.lblTorrent);
             this.Controls.Add(this.panel1);
@@ -188,12 +196,12 @@
 
         private System.Windows.Forms.ListView listResults;
         private System.Windows.Forms.ColumnHeader columnHeader2;
-        private System.Windows.Forms.ColumnHeader columnHeader3;
         private System.Windows.Forms.RichTextBox txtContents;
         private System.Windows.Forms.Panel panel1;
         private System.Windows.Forms.Label lblTorrent;
         private System.Windows.Forms.Label label1;
-        private System.Windows.Forms.Button button1;
-        private System.Windows.Forms.Button button2;
+        private System.Windows.Forms.Label lblHash;
+        private System.Windows.Forms.Button btnDownload;
+        private System.Windows.Forms.Button btnCopyHash;
     }
 }