+++ Torrentia/DetailsForm.cs
@@ -15,7 +15,7 @@ namespace Torrentia
     {
         SearchResult _result;
         TorrentzScraper _scraper = new TorrentzScraper();
-        string _magnet = string.Empty;
+        //string _magnet = string.Empty;
 
         public DetailsForm(SearchResult sr)
         {
@@ -25,52 +25,53 @@ namespace Torrentia
             _result = sr;
         }
 
-        private void OpenTorrentLink()
-        {
-            if (listResults.SelectedIndices.Count == 1)
-            {
-                try
-                {
-                    Process.Start(_result.Links[listResults.SelectedIndices[0]]);
-                }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(ex.Message, "Torrentia", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                }
-            }
-        }
+        //private void OpenTorrentLink()
+        //{
+        //    if (listResults.SelectedIndices.Count == 1)
+        //    {
+        //        try
+        //        {
+        //            Process.Start(_result.Links[listResults.SelectedIndices[0]]);
+        //        }
+        //        catch (Exception ex)
+        //        {
+        //            MessageBox.Show(ex.Message, "Torrentia", MessageBoxButtons.OK, MessageBoxIcon.Information);
+        //        }
+        //    }
+        //}
 
-        private void CopyMagnetURI()
-        {
-            if (listResults.SelectedIndices.Count == 1)
-            {
-                try
-                {
-                    _magnet = _scraper.GetMagnetUri(_result.Links[listResults.SelectedIndices[0]]);
-                    Clipboard.SetText(_magnet);
-                }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(ex.Message);
-                    _magnet = string.Empty;
-                }
-            }
-        }
+        //private void CopyMagnetURI()
+        //{
+        //    if (listResults.SelectedIndices.Count == 1)
+        //    {
+        //        try
+        //        {
+        //            _magnet = _scraper.GetMagnetUri(_result.Links[listResults.SelectedIndices[0]]);
+        //            Clipboard.SetText(_magnet);
+        //        }
+        //        catch (Exception ex)
+        //        {
+        //            MessageBox.Show(ex.Message);
+        //            _magnet = string.Empty;
+        //        }
+        //    }
+        //}
 
         private void DetailsForm_Load(object sender, EventArgs e)
         {
             lblTorrent.Text = _result.Verified + " " + _result.Title;
+            lblHash.Text += _result.InfoHash;
             txtContents.Text = _result.Contents.Trim();
 
-            int counter = 0;
+            //int counter = 0;
             foreach (string x in _result.Trackers)
             {
                 ListViewItem item = new ListViewItem(x);
                 
-                if (counter <= _result.Dates.Count - 1)
-                {
-                    item.SubItems.Add(_result.Dates[counter++]);
-                }
+                //if (counter <= _result.Dates.Count - 1)
+                //{
+                //    item.SubItems.Add(_result.Dates[counter++]);
+                //}
 
                 listResults.Items.Add(item);
             }
@@ -81,42 +82,72 @@ namespace Torrentia
 
         private void listResults_MouseDoubleClick(object sender, MouseEventArgs e)
         {
-            if (listResults.SelectedIndices.Count == 1)
-            {
-                LaunchTorrentClient();
-            }
+            //if (listResults.SelectedIndices.Count == 1)
+            //{
+            //    LaunchTorrentClient();
+            //}
         }
 
         private void LaunchTorrentClient()
         {
+            //try
+            //{
+            //    _magnet = _scraper.GetMagnetUri(_result.Links[listResults.SelectedIndices[0]]);
+            //}
+            //catch (Exception ex)
+            //{
+            //    MessageBox.Show(ex.Message, "Torrentia", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            //    _magnet = string.Empty;
+            //}
+
+            //if (!string.IsNullOrEmpty(_magnet))
+            //{
+            //    try
+            //    {
+            //        Process.Start(_magnet);
+            //    }
+            //    catch { }
+            //}
+        }
+
+        private string GenerateMagnetLink()
+        {
+            string result = "magnet:?xt=urn:btih:" + _result.InfoHash;
+
+            foreach (string x in _result.Trackers)
+            {
+                result += string.Format("&tr={0}", x.Replace("/", "%2F").Replace(":", "%3A"));
+            }
+
+            return result;
+        }
+
+        private void Download()
+        {
             try
             {
-                _magnet = _scraper.GetMagnetUri(_result.Links[listResults.SelectedIndices[0]]);
+                Process.Start(GenerateMagnetLink());
             }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message, "Torrentia", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                _magnet = string.Empty;
-            }
-
-            if (!string.IsNullOrEmpty(_magnet))
-            {
-                try
-                {
-                    Process.Start(_magnet);
-                }
-                catch { }
-            }
+            catch { }
         }
 
-        private void button1_Click(object sender, EventArgs e)
+        private void CopyHash()
         {
-            OpenTorrentLink();
+            try
+            {
+                Clipboard.SetText(lblHash.Text.Replace("Hash: ", string.Empty), TextDataFormat.Text);
+            }
+            catch { }
         }
 
-        private void button2_Click(object sender, EventArgs e)
+        private void btnCopyHash_Click(object sender, EventArgs e)
         {
-            CopyMagnetURI();
+            CopyHash();
+        }
+
+        private void btnDownload_Click(object sender, EventArgs e)
+        {
+            Download();
         }
     }
 }